FLASK/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
asttokens==2.4.1
async-timeout==4.0.1
attrs==21.2.0
Babel==2.8.0
bcc==0.18.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.3
bs4==0.0.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.6
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
comm==0.2.1
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
distro==1.7.0
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
frozenlist==1.2.0
gpg==1.16.0
greenlet==3.0.3
gunicorn==21.2.0
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
ipykernel==6.29.3
ipython==8.22.1
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.4
matplotlib==3.8.3
matplotlib-inline==0.1.6
more-itertools==8.10.0
moviepy==1.0.3
multidict==5.1.0
mysql-connector-python==8.3.0
natsort==8.0.2
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==9.0.1
platformdirs==4.2.0
pop-transition==1.1.2
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==3.12.4
proton-core==0.1.16
proton-keyring-linux==0.0.1
proton-keyring-linux-secretservice==0.0.1
proton-vpn-api-core==0.20.3
proton-vpn-connection==0.11.3
proton-vpn-gtk-app==4.1.10
proton-vpn-killswitch==0.2.0
proton-vpn-killswitch-network-manager==0.2.0
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.3.3
proton-vpn-network-manager-openvpn==0.0.4
proton-vpn-session==0.6.5
psutil==5.9.8
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
systemd-python==234
tabulate==0.8.9
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
